refactor: Move aroung classes, and rename objects
