+++ cardpresent/src/main/java/com/staxpayments/android/chipdna/ChipDnaUtils.kt
@@ -10,7 +10,7 @@ import com.staxpayments.sdk.data.MobileReader
 import com.staxpayments.sdk.data.TransactionRequest
 import com.staxpayments.sdk.data.TransactionUpdate
 import com.staxpayments.sdk.data.UserNotification
-import com.staxpayments.api.models.MobileReaderConnectionStatus
+import com.staxpayments.sdk.MobileReaderConnectionStatus
 import com.staxpayments.api.models.Transaction
 import java.text.SimpleDateFormat
 import java.util.Date
