+++ cardpresent/src/main/java/com/staxpayments/sdk/MobileReaderConnectionStatusListener.kt
@@ -1,7 +1,5 @@
 package com.staxpayments.sdk
 
-import com.staxpayments.api.models.MobileReaderConnectionStatus
-
 interface MobileReaderConnectionStatusListener {
 
     /**
